MLhelp/requirements.txt
adagio==0.2.6
aiofiles==23.2.1
alembic==1.14.0
annotated-types==0.7.0
ansi2html==1.9.2
antlr4-python3-runtime==4.11.1
anyio==4.7.0
aplr==10.8.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.35.76
botocore==1.35.76
cachetools==5.5.0
catboost==1.2.7
category-encoders==2.6.4
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
colorlog==6.9.0
colour==0.1.5
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
daal==2025.0.0
dacite==1.8.1
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-testing-stub==0.0.2
dash_auth==2.3.0
dash_cytoscape==1.0.2
dask==2024.12.0
dask-expr==1.1.20
databricks-sdk==0.38.0
debugpy==1.8.9
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.9
distributed==2024.12.0
distro==1.9.0
Django==5.1.1
djangorestframework==3.15.2
docker==7.1.0
dtreeviz==2.2.2
dynaconf==3.2.6
evidently==0.4.40
exceptiongroup==1.2.2
executing==2.1.0
explainerdashboard==0.4.7
fairlearn==0.7.0
Faker==33.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
ffmpy==0.4.0
filelock==3.16.1
Flask==2.3.3
Flask-WTF==1.2.2
flask_simplelogin==0.2.0
fonttools==4.54.1
fs==2.4.16
fsspec==2024.9.0
fugue==0.8.7
fugue-sql-antlr==0.2.2
future==1.0.0
gevent==24.11.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
gradio==5.8.0
gradio_client==1.5.1
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.26.5
hyperopt==0.2.7
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.12.4
importlib_metadata==8.5.0
iniconfig==2.0.0
interpret==0.6.7
interpret-core==0.6.7
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
iterative-telemetry==0.0.9
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-dash==0.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.7
kmodes==0.12.2
lightgbm==4.5.0
litestar==2.13.0
litestar-htmx==0.4.1
llvmlite==0.43.0
locket==1.0.0
lxml==5.3.0
m2cgen==0.10.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.18.0
mlflow-skinny==2.18.0
mlxtend==0.23.3
moto==4.2.14
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multimethod==1.12
multiprocess==0.70.17
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
optuna==4.1.0
optuna-integration==4.1.0
orjson==3.10.12
outcome==1.3.0.post0
oyaml==1.0
packaging==24.2
pandas==2.1.4
parso==0.8.4
partd==1.4.2
patsy==1.0.1
percy==2.0.2
phik==0.12.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
plotly-resampler==0.10.0
pluggy==1.5.0
pmdarima==2.0.4
polyfactory==2.18.1
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
py4j==0.10.9.7
pyaml==24.9.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycaret==3.3.2
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyod==2.0.2
pyOpenSSL==24.3.0
pyparsing==3.2.0
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyWavelets==1.8.0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
qpd==0.4.4
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
responses==0.25.3
retrying==1.3.4
rich==13.9.4
rich-click==1.8.5
rpds-py==0.22.3
rsa==4.9
ruff==0.8.2
s3transfer==0.10.4
safehttpx==0.1.6
SALib==1.5.1
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-learn-intelex==2025.0.0
scikit-optimize==0.10.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
selenium==4.2.0
semantic-version==2.10.0
shap==0.44.1
shellingham==1.5.4
six==1.16.0
sktime==0.26.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlglot==25.34.0
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.3
statsforecast==1.5.0
statsmodels==0.14.4
tbats==1.1.3
tbb==2022.0.0
tblib==3.0.0
tcmlib==1.2.0
tenacity==9.0.0
threadpoolctl==3.5.0
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triad==0.9.8
trio==0.27.0
trio-websocket==0.11.1
tsdownsample==0.1.3
typeguard==4.4.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
umap-learn==0.5.7
urllib3==1.26.20
urllib3-secure-extra==0.1.0
uuid6==2024.7.10
uvicorn==0.32.1
visions==0.7.6
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==2.3.8
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.0
wsproto==1.2.0
WTForms==3.2.1
xgboost==2.1.3
xmltodict==0.14.2
xxhash==3.5.0
ydata-profiling==4.12.1
yellowbrick==1.5
zict==3.0.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
